+++ Harmony.PatchExtensions/TranspilerConfig.cs
@@ -4,13 +4,16 @@ namespace HarmonyLib.PatchExtensions;
 
 internal class TranspilerConfig
 {
-    public TranspilerConfig(AT type, string targetMember, MethodInfo patchMethod, uint occurrence, uint startIndex)
+    public TranspilerConfig(AT type, string targetMember, MethodInfo patchMethod, uint occurrence, uint startIndex,
+        uint argIndex
+    )
     {
         Type = type;
         TargetMember = targetMember;
         PatchMethod = patchMethod;
         Occurrence = occurrence;
         StartIndex = startIndex;
+        ArgIndex = argIndex;
     }
     
     public AT Type;
@@ -18,4 +21,5 @@ internal class TranspilerConfig
     public MethodInfo PatchMethod;
     public uint Occurrence;
     public uint StartIndex;
+    public uint ArgIndex;
 }